Added initial files and outline
preprocessTaskAndKinematicData script/function do some low-level kinematic preprocessing (smoothing, upsampling) and pulls out trial labels for the choking task.

Target.m and ChangeTarget.m are classes used in the task for the targets and should be in the path for running this stuff (ChangeTarget is only used in the focus task).